CITY-37 lavet UserService interface jf CBD
