+++ src/main/java/com/example/citycutbackend/user/UserController.java
@@ -12,10 +12,10 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class UserController {
 
-    private UserService userService;
+    private UserServiceImpl userService;
 
 
-    public UserController(UserService userService) {
+    public UserController(UserServiceImpl userService) {
         this.userService = userService;
     }
